+++ src/engine/GameEngine.ts
@@ -32,6 +32,11 @@ export interface GameEngineOptions {
     imageLoader?: ImageLoader;
 
     /**
+     * The render system to use for rendering the scene. If not provided, a default render system will be used.
+     */
+    renderSystem?: (renderer: Renderer, imageLoader: ImageLoader) => RenderSystem;
+
+    /**
      * The factory function to use for adding additional render systems.
      */
     additionalRenderSystems?: (renderer: Renderer, imageLoader: ImageLoader) => RenderSystem[];
@@ -76,9 +81,12 @@ export class GameEngine {
 
         this.imageLoader = config.imageLoader ?? new DOMImageLoader();
 
-        this.renderSystems = [new RenderSystem(this.renderer, this.imageLoader)].concat(
+        const renderSystem = config.renderSystem ? config.renderSystem(this.renderer, this.imageLoader) : new RenderSystem(this.renderer, this.imageLoader);
+
+        this.renderSystems = [renderSystem].concat(
             config.additionalRenderSystems ? config.additionalRenderSystems(this.renderer, this.imageLoader) : []
         );
+
         this.physicsSystem = new PhysicsSystem(this._physx);
         this.hierarchySystem = new HierarchySystem();
         this.inputSystem = new InputSystem(this._inputs);